requirements.txt
streamlit
pandas
Pillow
matplotlib
seaborn
